Use normal textbox for ID/Hash
Reasoning: This field is most often used by pasting in a value to search, and the dropdown gets in the way of the button. Also, even if you *are* able to press the button, it won't use the updated value because it doesn't get set until pressing enter or clicking elsewhere outside the field.